Adding Response Formatter

+++ src/main/java/com/firefly/myapp/controller/MessageController.java
@@ -2,6 +2,7 @@ package com.firefly.myapp.controller;
 
 import com.firefly.myapp.model.Messages;
 import com.firefly.myapp.service.MessageService;
+import com.firefly.myapp.utils.ResponseFormatter;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
@@ -14,31 +15,48 @@ public class MessageController {
     MessageService messageService;
 
     @GetMapping
-    private List<Messages> index() {
-        return messageService.findAllMessages();
+    private ResponseFormatter<List<Messages>> index() {
+        List<Messages> messages = messageService.findAllMessages();
+        return new ResponseFormatter<>(true, "Get all messages", messages);
     }
 
     @GetMapping("/{id}")
-    private Messages show(@PathVariable("id") int id) {
-        return messageService.findMessageById(id);
-    }
-
-    @DeleteMapping("/{id}")
-    private void delete(@PathVariable("id") int id) {
-        messageService.delete(id);
+    private ResponseFormatter<Messages> show(@PathVariable("id") int id) {
+        Messages message = messageService.findMessageById(id);
+        if (message != null) {
+            return new ResponseFormatter<>(message);
+        } else{
+            return new ResponseFormatter<>(false, "Message Not Found");
+        }
     }
 
     @PostMapping
-    private int store(@RequestBody Messages messages)
+    private ResponseFormatter<Void> store(@RequestBody Messages messages)
     {
         messageService.saveOrUpdate(messages);
-        return messages.getId();
+        return new ResponseFormatter<>(true, "Message Has Been Created");
     }
 
     @PutMapping("/{id}")
-    private int update(@RequestBody Messages messages, @PathVariable("id") int id)
+    private ResponseFormatter<Void> update(@RequestBody Messages messages, @PathVariable("id") int id)
     {
-        messageService.update(messages, id);
-        return messages.getId();
+        Messages message = messageService.findMessageById(id);
+        if (message != null) {
+            messageService.update(messages, id);
+            return new ResponseFormatter<>(true, "Message Has Been Updated");
+        } else {
+            return new ResponseFormatter<>(false, "Message Not Found");
+        }
+    }
+
+    @DeleteMapping("/{id}")
+    private ResponseFormatter<Void> delete(@PathVariable("id") int id) {
+        Messages message = messageService.findMessageById(id);
+        if (message != null) {
+            messageService.delete(id);
+            return new ResponseFormatter<>(true, "Message Has Been Deleted");
+        } else {
+            return new ResponseFormatter<>(false, "Message Not Found");
+        }
     }
 }
